+++ podpointclient/__init__.py
@@ -2,20 +2,22 @@
 
 from .client import PodPointClient
 from .domain import (
-    AccountCapability, BoostState, CapabilitySupport, ChargeSession,
+    AccountCapability, BasicChargingMode, BoostState, CapabilitySupport, ChargeSession,
     ChargerCapability, ChargerDomain, ChargerIdentityError, ChargerRef,
     ChargerSource, ChargerState, NormalizedStateValue, StateValue,
     boost_state_from_home, boost_state_from_legacy, charger_ref_from_charger,
     charger_ref_from_pod, charge_session_from_home, charge_session_from_legacy,
-    normalize_state,
+    normalize_state, reconcile_charge_sessions,
 )
 from .errors import UnsupportedCapabilityError
 
 __all__ = [
-    "AccountCapability", "BoostState", "CapabilitySupport", "ChargeSession",
+    "AccountCapability", "BasicChargingMode", "BoostState", "CapabilitySupport",
+    "ChargeSession",
     "ChargerCapability", "ChargerDomain", "ChargerIdentityError", "ChargerRef",
     "ChargerSource", "ChargerState", "NormalizedStateValue", "PodPointClient",
     "StateValue", "UnsupportedCapabilityError", "boost_state_from_home",
     "boost_state_from_legacy", "charger_ref_from_charger", "charger_ref_from_pod",
     "charge_session_from_home", "charge_session_from_legacy", "normalize_state",
+    "reconcile_charge_sessions",
 ]